+++ spec/feature_spec.rb
@@ -1,3 +1,5 @@
+require 'my_helper'
+
 feature 'Playability feature Test' do
 
   context 'single player' do
@@ -12,8 +14,7 @@ feature 'Playability feature Test' do
       click_button 'Play!'
       expect(page).to have_content 'Pietro chose rock Watson chose scissors'
       expect(page).to have_content 'Pietro won!'
-      expect(page).to have_link 'Play Again...'
-      expect(page).to have_link 'Go To Two Player Game'
+      bottom_expectations
     end
 
     it 'game version against Ai - draw' do
@@ -26,8 +27,7 @@ feature 'Playability feature Test' do
       click_button 'Play!'
       expect(page).to have_content 'Pietro chose scissors Watson chose scissors'
       expect(page).to have_content 'It was a draw'
-      expect(page).to have_link 'Play Again...'
-      expect(page).to have_link 'Go To Two Player Game'
+      bottom_expectations
     end
 
   end
@@ -46,8 +46,7 @@ feature 'Playability feature Test' do
       click_button 'Play!'
       expect(page).to have_content 'Pietro chose rock Joanna chose paper'
       expect(page).to have_content 'Joanna won!'
-      expect(page).to have_link 'Play Again...'
-      expect(page).to have_link 'Go To Two Player Game'
+      bottom_expectations
     end
 
     it 'game version - draw' do
@@ -61,8 +60,7 @@ feature 'Playability feature Test' do
       select "Scissors", :from => "weapon2"
       click_button 'Play!'
       expect(page).to have_content 'It was a draw'
-      expect(page).to have_link 'Play Again...'
-      expect(page).to have_link 'Go To Two Player Game'
+      bottom_expectations
     end
 
   end
