+++ spec/spec_helper.rb
@@ -3,8 +3,6 @@ require 'simplecov'
 require 'simplecov-console'
 require_relative '../app.rb'
 
-Capybara.app = RPSWeb
-
 SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new([
   SimpleCov::Formatter::Console,
   # Want a nice code coverage website? Uncomment this next line!
@@ -17,3 +15,5 @@ RSpec.configure do |config|
 
   end
 end
+
+Capybara.app = RPSWeb
