+++ views/draw.erb
@@ -1,4 +1,3 @@
 <%= erb(:scoreboard) %>
-
 <p class="text-info">It was a draw</p>
-<a href="/" class="btn btn-success">Play Again...</a>
+<h3><%= erb(:footer) %></h3>
