+++ views/footer.erb
@@ -1,0 +1,2 @@
+<a href="/" class="btn btn-primary">Play Again...</a>
+<a href="/two_player" class="btn btn-primary">Play Two Players...</a>
