+++ views/index.erb
@@ -1,3 +1,6 @@
+<div class="row text-center" style="margin-top: 20px">
+  <a href='/two_player' class="btn btn-primary">Go To Two Player Game</a>
+</div>
 
 <%= erb(:single_player) %>
