+++ views/scoreboard.erb
@@ -1,5 +1,5 @@
-<p class="text-info"><%= @game.player_1.name %> chose <%= @game.player_1.choice %></p>
+<h3 class="text-info"><%= @game.player_1.name %> chose <%= @game.player_1.choice %></p>
 <br>
 
-<p class="text-info"><%= @game.player_2.name %> chose <%= @game.player_2.choice %></p>
+<h3 class="text-info"><%= @game.player_2.name %> chose <%= @game.player_2.choice %></p>
 <br>
