+++ views/single_player.erb
@@ -1,3 +1,4 @@
-<div class="page-header">
-    <h1>Rock Paper Scissors <small>Single Player</small></h1>
+<div class="header">
+    <h1>Rock Paper Scissors</h1>
+    <h1><small>Single Player</small></h1>
 </div>
