+++ views/two_player.erb
@@ -1,5 +1,10 @@
-<div class="page-header">
-    <h1>Rock Paper Scissors <small>Two Players</small></h1>
+<div class="row text-center" style="margin-top: 20px">
+  <a href='/' class="btn btn-primary">Go To Single Player Game</a>
+</div>
+
+<div class="header">
+    <h1>Rock Paper Scissors</h1>
+    <h1><small>Two Players</small></h1>
 </div>
 
 <form name="rpsForm" action="/play" method="post" onsubmit="return validateForm()" id="rpsForm">
@@ -31,6 +36,4 @@
         </p>
 
         <input type="submit" class="btn btn-primary" value="Play!">
-
-
     </form>
