+++ views/victory.erb
@@ -1,3 +1,3 @@
-<%= erb(:scoreboard) %>
-<p class="text-info"><%= @game.who_won %> won!</p>
-<a href="/" class="btn btn-success">Play Again...</a>
+<h3><%= erb(:scoreboard) %></h3>
+<h3 class="text-info"><%= @game.who_won %> won!</h3>
+<h3><%= erb(:footer) %></h3>
